refactor(grid): simplify internals by extracting concerns to focused modules
- Extract ShellController to dedicated shell-controller.ts module, removing ~320 lines of duplicated code from shell.ts and rewiring grid.ts imports
- De-duplicate shell DOM builders by extracting shared buildShellOptions helper, eliminating repeated option-building logic in buildGridDOMIntoElement and rebuildShellDOM
- Simplify compileTemplate in sanitize.ts by returning early for blocked templates instead of wrapping in a redundant forceBlank closure

Bundle impact: -473 bytes raw, -146 bytes gzip (160,061 / 45,326)
