+++ libs/grid/src/lib/core/internal/shell.ts
@@ -10,20 +10,11 @@
 import type {
   HeaderContentDefinition,
   IconValue,
-  InternalGrid,
   ShellConfig,
   ToolbarContentDefinition,
   ToolPanelDefinition,
 } from '../types';
-import {
-  HEADER_CONTENT_DUPLICATE,
-  NO_TOOL_PANELS,
-  TOOL_PANEL_DUPLICATE,
-  TOOL_PANEL_MISSING_ATTR,
-  TOOL_PANEL_NOT_FOUND,
-  TOOLBAR_CONTENT_DUPLICATE,
-  warnDiagnostic,
-} from './diagnostics';
+import { TOOL_PANEL_MISSING_ATTR, warnDiagnostic } from './diagnostics';
 import { escapeHtml, sanitizeHTML } from './sanitize';
 
 // #region Types & State
@@ -960,322 +951,6 @@ export function cleanupShellState(state: ShellState): void {
 }
 // #endregion
 
-// #region ShellController
-/**
- * Controller interface for managing shell/tool panel behavior.
- */
-export interface ShellController {
-  /** Whether the shell has been initialized */
-  readonly isInitialized: boolean;
-  /** Set the initialized state */
-  setInitialized(value: boolean): void;
-  /** Whether the tool panel is currently open */
-  readonly isPanelOpen: boolean;
-  /** Get IDs of expanded accordion sections */
-  readonly expandedSections: string[];
-  /** Open the tool panel */
-  openToolPanel(): void;
-  /** Close the tool panel */
-  closeToolPanel(): void;
-  /** Toggle the tool panel */
-  toggleToolPanel(): void;
-  /** Toggle an accordion section */
-  toggleToolPanelSection(sectionId: string): void;
-  /** Get registered tool panels */
-  getToolPanels(): ToolPanelDefinition[];
-  /** Register a tool panel */
-  registerToolPanel(panel: ToolPanelDefinition): void;
-  /** Unregister a tool panel */
-  unregisterToolPanel(panelId: string): void;
-  /** Get registered header contents */
-  getHeaderContents(): HeaderContentDefinition[];
-  /** Register header content */
-  registerHeaderContent(content: HeaderContentDefinition): void;
-  /** Unregister header content */
-  unregisterHeaderContent(contentId: string): void;
-  /** Get all registered toolbar contents */
-  getToolbarContents(): ToolbarContentDefinition[];
-  /** Register toolbar content */
-  registerToolbarContent(content: ToolbarContentDefinition): void;
-  /** Unregister toolbar content */
-  unregisterToolbarContent(contentId: string): void;
-}
-
-/**
- * Create a ShellController instance.
- * The controller encapsulates all tool panel orchestration logic.
- */
-export function createShellController(state: ShellState, grid: InternalGrid): ShellController {
-  let initialized = false;
-
-  const controller: ShellController = {
-    get isInitialized() {
-      return initialized;
-    },
-    setInitialized(value: boolean) {
-      initialized = value;
-    },
-
-    get isPanelOpen() {
-      return state.isPanelOpen;
-    },
-
-    get expandedSections() {
-      return [...state.expandedSections];
-    },
-
-    openToolPanel() {
-      if (state.isPanelOpen) return;
-      if (state.toolPanels.size === 0) {
-        warnDiagnostic(NO_TOOL_PANELS, 'No tool panels registered', grid.id);
-        return;
-      }
-
-      state.isPanelOpen = true;
-
-      // Auto-expand first section if none expanded
-      if (state.expandedSections.size === 0 && state.toolPanels.size > 0) {
-        const sortedPanels = [...state.toolPanels.values()].sort((a, b) => (a.order ?? 100) - (b.order ?? 100));
-        const firstPanel = sortedPanels[0];
-        if (firstPanel) {
-          state.expandedSections.add(firstPanel.id);
-        }
-      }
-
-      // Update UI
-      const shadow = grid._renderRoot;
-      updateToolbarActiveStates(shadow, state);
-      updatePanelState(shadow, state);
-
-      // Render accordion sections
-      renderPanelContent(shadow, state, grid._accordionIcons);
-
-      // Emit event
-      grid._emit('tool-panel-open', { sections: controller.expandedSections });
-    },
-
-    closeToolPanel() {
-      if (!state.isPanelOpen) return;
-
-      // Clean up all panel content
-      for (const cleanup of state.panelCleanups.values()) {
-        cleanup();
-      }
-      state.panelCleanups.clear();
-
-      // Call onClose for all panels
-      for (const panel of state.toolPanels.values()) {
-        panel.onClose?.();
-      }
-
-      state.isPanelOpen = false;
-
-      // Update UI
-      const shadow = grid._renderRoot;
-      updateToolbarActiveStates(shadow, state);
-      updatePanelState(shadow, state);
-
-      // Emit event
-      grid._emit('tool-panel-close', {});
-    },
-
-    toggleToolPanel() {
-      if (state.isPanelOpen) {
-        controller.closeToolPanel();
-      } else {
-        controller.openToolPanel();
-      }
-    },
-
-    toggleToolPanelSection(sectionId: string) {
-      const panel = state.toolPanels.get(sectionId);
-      if (!panel) {
-        warnDiagnostic(TOOL_PANEL_NOT_FOUND, `Tool panel section "${sectionId}" not found`, grid.id);
-        return;
-      }
-
-      // Don't allow toggling when there's only one panel (it should stay expanded)
-      if (state.toolPanels.size === 1) {
-        return;
-      }
-
-      const shadow = grid._renderRoot;
-      const isExpanded = state.expandedSections.has(sectionId);
-
-      if (isExpanded) {
-        // Collapsing current section
-        const cleanup = state.panelCleanups.get(sectionId);
-        if (cleanup) {
-          cleanup();
-          state.panelCleanups.delete(sectionId);
-        }
-        panel.onClose?.();
-        state.expandedSections.delete(sectionId);
-        updateAccordionSectionState(shadow, sectionId, false);
-      } else {
-        // Expanding - first collapse all others (exclusive accordion)
-        for (const [otherId, otherPanel] of state.toolPanels) {
-          if (otherId !== sectionId && state.expandedSections.has(otherId)) {
-            const cleanup = state.panelCleanups.get(otherId);
-            if (cleanup) {
-              cleanup();
-              state.panelCleanups.delete(otherId);
-            }
-            otherPanel.onClose?.();
-            state.expandedSections.delete(otherId);
-            updateAccordionSectionState(shadow, otherId, false);
-            // Clear content of collapsed section
-            const contentEl = shadow.querySelector(`[data-section="${otherId}"] .tbw-accordion-content`);
-            if (contentEl) contentEl.innerHTML = '';
-          }
-        }
-        // Now expand the target section
-        state.expandedSections.add(sectionId);
-        updateAccordionSectionState(shadow, sectionId, true);
-        renderAccordionSectionContent(shadow, state, sectionId);
-      }
-
-      // Emit event
-      grid._emit('tool-panel-section-toggle', { id: sectionId, expanded: !isExpanded });
-    },
-
-    getToolPanels() {
-      return [...state.toolPanels.values()];
-    },
-
-    registerToolPanel(panel: ToolPanelDefinition) {
-      if (state.toolPanels.has(panel.id)) {
-        warnDiagnostic(TOOL_PANEL_DUPLICATE, `Tool panel "${panel.id}" already registered`, grid.id);
-        return;
-      }
-      state.toolPanels.set(panel.id, panel);
-
-      if (initialized) {
-        grid.refreshShellHeader?.();
-      }
-    },
-
-    unregisterToolPanel(panelId: string) {
-      // Close panel if open and this section is expanded
-      if (state.expandedSections.has(panelId)) {
-        const cleanup = state.panelCleanups.get(panelId);
-        if (cleanup) {
-          cleanup();
-          state.panelCleanups.delete(panelId);
-        }
-        state.expandedSections.delete(panelId);
-      }
-
-      state.toolPanels.delete(panelId);
-
-      if (initialized) {
-        grid.refreshShellHeader?.();
-      }
-    },
-
-    getHeaderContents() {
-      return [...state.headerContents.values()];
-    },
-
-    registerHeaderContent(content: HeaderContentDefinition) {
-      if (state.headerContents.has(content.id)) {
-        warnDiagnostic(HEADER_CONTENT_DUPLICATE, `Header content "${content.id}" already registered`, grid.id);
-        return;
-      }
-      state.headerContents.set(content.id, content);
-
-      if (initialized) {
-        renderHeaderContent(grid._renderRoot, state);
-      }
-    },
-
-    unregisterHeaderContent(contentId: string) {
-      // Clean up
-      const cleanup = state.headerContentCleanups.get(contentId);
-      if (cleanup) {
-        cleanup();
-        state.headerContentCleanups.delete(contentId);
-      }
-
-      // Call onDestroy
-      const content = state.headerContents.get(contentId);
-      content?.onDestroy?.();
-
-      state.headerContents.delete(contentId);
-
-      // Remove DOM element
-      const el = grid._renderRoot.querySelector(`[data-header-content="${contentId}"]`);
-      el?.remove();
-    },
-
-    getToolbarContents() {
-      return [...state.toolbarContents.values()].sort((a, b) => (a.order ?? 0) - (b.order ?? 0));
-    },
-
-    registerToolbarContent(content: ToolbarContentDefinition) {
-      if (state.toolbarContents.has(content.id)) {
-        warnDiagnostic(TOOLBAR_CONTENT_DUPLICATE, `Toolbar content "${content.id}" already registered`, grid.id);
-        return;
-      }
-      state.toolbarContents.set(content.id, content);
-
-      if (initialized) {
-        grid.refreshShellHeader?.();
-      }
-    },
-
-    unregisterToolbarContent(contentId: string) {
-      // Clean up
-      const cleanup = state.toolbarContentCleanups.get(contentId);
-      if (cleanup) {
-        cleanup();
-        state.toolbarContentCleanups.delete(contentId);
-      }
-
-      // Call onDestroy if defined
-      const content = state.toolbarContents.get(contentId);
-      if (content?.onDestroy) {
-        content.onDestroy();
-      }
-
-      state.toolbarContents.delete(contentId);
-
-      if (initialized) {
-        grid.refreshShellHeader?.();
-      }
-    },
-  };
-
-  return controller;
-}
-
-/**
- * Update accordion section visual state.
- */
-function updateAccordionSectionState(renderRoot: Element, sectionId: string, expanded: boolean): void {
-  const section = renderRoot.querySelector(`[data-section="${sectionId}"]`);
-  if (section) {
-    section.classList.toggle('expanded', expanded);
-  }
-}
-
-/**
- * Render content for a single accordion section.
- */
-function renderAccordionSectionContent(renderRoot: Element, state: ShellState, sectionId: string): void {
-  const panel = state.toolPanels.get(sectionId);
-  if (!panel?.render) return;
-
-  const contentEl = renderRoot.querySelector(`[data-section="${sectionId}"] .tbw-accordion-content`);
-  if (!contentEl) return;
-
-  const cleanup = panel.render(contentEl as HTMLElement);
-  if (cleanup) {
-    state.panelCleanups.set(sectionId, cleanup);
-  }
-}
-// #endregion
-
 // #region Grid HTML Templates
 /**
  * Core grid content HTML template.
@@ -1314,12 +989,59 @@ import {
 } from './dom-builder';
 
 /**
+ * Build ShellHeaderOptions and ShellBodyOptions from shell config and runtime state.
+ * Shared by buildGridDOMIntoElement and rebuildShellDOM to avoid duplication.
+ */
+function buildShellOptions(
+  shellConfig: ShellConfig,
+  runtimeState: { isPanelOpen: boolean; expandedSections: Set<string> },
+  icons?: { toolPanel?: IconValue; expand?: IconValue; collapse?: IconValue },
+): { headerOptions: ShellHeaderOptions; bodyOptions: ShellBodyOptions } {
+  const toolPanelIcon = icons?.toolPanel !== undefined ? iconToString(icons.toolPanel) : undefined;
+  const expandIcon = icons?.expand !== undefined ? iconToString(icons.expand) : undefined;
+  const collapseIcon = icons?.collapse !== undefined ? iconToString(icons.collapse) : undefined;
+
+  const allContents = shellConfig.header?.toolbarContents ?? [];
+  const sortedContents = [...allContents].sort((a, b) => (a.order ?? 0) - (b.order ?? 0));
+
+  const allPanels = shellConfig.toolPanels ?? [];
+  const sortedPanels = [...allPanels].sort((a, b) => (a.order ?? 100) - (b.order ?? 100));
+
+  return {
+    headerOptions: {
+      title: shellConfig.header?.title ?? undefined,
+      hasPanels: sortedPanels.length > 0,
+      isPanelOpen: runtimeState.isPanelOpen,
+      toolPanelIcon,
+      configButtons: sortedContents.map((c) => ({
+        id: c.id,
+        hasElement: false,
+        hasRender: !!c.render,
+      })),
+      apiButtons: [],
+    },
+    bodyOptions: {
+      position: shellConfig.toolPanel?.position ?? 'right',
+      isPanelOpen: runtimeState.isPanelOpen,
+      expandIcon,
+      collapseIcon,
+      panels: sortedPanels.map((p) => ({
+        id: p.id,
+        title: p.title,
+        icon: iconToString(p.icon),
+        isExpanded: runtimeState.expandedSections.has(p.id),
+      })),
+    },
+  };
+}
+
+/**
  * Build the complete grid DOM structure using direct DOM construction.
  * This is 2-3x faster than innerHTML for initial render.
  *
  * @param renderRoot - The element to render into (will be cleared)
  * @param shellConfig - Shell configuration
- * @param state - Shell state
+ * @param runtimeState - Runtime shell state
  * @param icons - Optional icons
  * @returns Whether shell is active (for post-render setup)
  */
@@ -1358,52 +1080,10 @@ export function buildGridDOMIntoElement(
   }
 
   if (hasShell) {
-    const toolPanelIcon = icons?.toolPanel !== undefined ? iconToString(icons.toolPanel) : undefined;
-    const expandIcon = icons?.expand !== undefined ? iconToString(icons.expand) : undefined;
-    const collapseIcon = icons?.collapse !== undefined ? iconToString(icons.collapse) : undefined;
-
-    // All toolbar contents now come from shellConfig (merged by ConfigManager)
-    const allContents = shellConfig?.header?.toolbarContents ?? [];
-    const sortedContents = [...allContents].sort((a, b) => (a.order ?? 0) - (b.order ?? 0));
-
-    // All panels now come from shellConfig (merged by ConfigManager)
-    const allPanels = shellConfig?.toolPanels ?? [];
-    const sortedPanels = [...allPanels].sort((a, b) => (a.order ?? 100) - (b.order ?? 100));
-
-    // Build header options
-    const headerOptions: ShellHeaderOptions = {
-      title: shellConfig?.header?.title ?? undefined,
-      hasPanels: sortedPanels.length > 0,
-      isPanelOpen: runtimeState.isPanelOpen,
-      toolPanelIcon,
-      // All contents are now in config (no more separate config vs API distinction for rendering)
-      configButtons: sortedContents.map((c) => ({
-        id: c.id,
-        hasElement: false,
-        hasRender: !!c.render,
-      })),
-      apiButtons: [], // No longer needed - all contents merged into configButtons
-    };
-
-    // Build body options
-    const bodyOptions: ShellBodyOptions = {
-      position: shellConfig?.toolPanel?.position ?? 'right',
-      isPanelOpen: runtimeState.isPanelOpen,
-      expandIcon,
-      collapseIcon,
-      panels: sortedPanels.map((p) => ({
-        id: p.id,
-        title: p.title,
-        icon: iconToString(p.icon),
-        isExpanded: runtimeState.expandedSections.has(p.id),
-      })),
-    };
-
-    // Build shell elements
+    const { headerOptions, bodyOptions } = buildShellOptions(shellConfig!, runtimeState, icons);
     const shellHeader = buildShellHeader(headerOptions);
     const shellBody = buildShellBody(bodyOptions);
 
-    // Build and append complete DOM
     const fragment = buildGridDOM({
       hasShell: true,
       shellHeader,
@@ -1411,7 +1091,6 @@ export function buildGridDOMIntoElement(
     });
     renderRoot.appendChild(fragment);
   } else {
-    // No shell - just grid content
     const fragment = buildGridDOM({ hasShell: false });
     renderRoot.appendChild(fragment);
   }
@@ -1457,43 +1136,7 @@ export function rebuildShellDOM(
   if (hasShell) {
     existingRoot.className = `${GridClasses.ROOT} has-shell`;
 
-    const toolPanelIcon = icons?.toolPanel !== undefined ? iconToString(icons.toolPanel) : undefined;
-    const expandIcon = icons?.expand !== undefined ? iconToString(icons.expand) : undefined;
-    const collapseIcon = icons?.collapse !== undefined ? iconToString(icons.collapse) : undefined;
-
-    const allContents = shellConfig?.header?.toolbarContents ?? [];
-    const sortedContents = [...allContents].sort((a, b) => (a.order ?? 0) - (b.order ?? 0));
-
-    const allPanels = shellConfig?.toolPanels ?? [];
-    const sortedPanels = [...allPanels].sort((a, b) => (a.order ?? 100) - (b.order ?? 100));
-
-    const headerOptions: ShellHeaderOptions = {
-      title: shellConfig?.header?.title ?? undefined,
-      hasPanels: sortedPanels.length > 0,
-      isPanelOpen: runtimeState.isPanelOpen,
-      toolPanelIcon,
-      configButtons: sortedContents.map((c) => ({
-        id: c.id,
-        hasElement: false,
-        hasRender: !!c.render,
-      })),
-      apiButtons: [],
-    };
-
-    const bodyOptions: ShellBodyOptions = {
-      position: shellConfig?.toolPanel?.position ?? 'right',
-      isPanelOpen: runtimeState.isPanelOpen,
-      expandIcon,
-      collapseIcon,
-      panels: sortedPanels.map((p) => ({
-        id: p.id,
-        title: p.title,
-        icon: iconToString(p.icon),
-        isExpanded: runtimeState.expandedSections.has(p.id),
-      })),
-    };
-
-    // Build new shell elements
+    const { headerOptions, bodyOptions } = buildShellOptions(shellConfig!, runtimeState, icons);
     const shellHeader = buildShellHeader(headerOptions);
     const shellBody = buildShellBody(bodyOptions);
